@jjhursey 2022-10-19
I'm not well versed in the add_procs requirements nor the UCX internals so I need the @open-mpi/ucx folks to help review to make sure this fits with their model.

Manual profiling of Open MPI down through the UCX PML isolated the performance problem.

I have a test case at the Gist below:
 * https://gist.github.com/jjhursey/ee2f611121a635a8edd59eb1dc8fe30e
 * Tested creating 4000 intercommunicators with the test case above
   * `Create 3` is the `MPI_Intercomm_create` call.
 * UCX 1.13.0 / MOFED 5.6-1.0.3.6

## TCP Baseline

```
shell$ mpirun -np 4 --mca pml ob1 --mca btl tcp,sm,self --host f8n07:2,f8n08:2 ./a.out 4000
iter count is 4000
Size of MPI_COMM_WORLD = 4
0/4) Create 0:   0.04 sec (  0.00 sec per iteration)
0/4) Create 1:   0.05 sec (  0.00 sec per iteration)
0/4) Create 2:   3.88 sec (  0.00 sec per iteration)
0/4) Create 3:   8.38 sec (  0.00 sec per iteration)
0/4) Total   :  12.35 sec (  0.00 sec per iteration)
3/4) Finalize:   0.14 sec
2/4) Finalize:   0.15 sec
1/4) Finalize:   0.38 sec
0/4) Finalize:   0.38 sec
```

## Without this patch

```
shell$ mpirun -np 4 --mca pml ucx --mca btl self --host f8n07:2,f8n08:2 ./a.out 4000
iter count is 4000
Size of MPI_COMM_WORLD = 4
0/4) Create 0:   0.03 sec (  0.00 sec per iteration)
0/4) Create 1:   0.04 sec (  0.00 sec per iteration)
0/4) Create 2:   0.30 sec (  0.00 sec per iteration)
0/4) Create 3:  57.97 sec (  0.01 sec per iteration)
0/4) Total   :  58.34 sec (  0.01 sec per iteration)
3/4) Finalize:  27.03 sec
2/4) Finalize:  27.57 sec
1/4) Finalize:  43.84 sec
0/4) Finalize:  44.22 sec
```

## With this patch

```
shell$ mpirun -np 4 --mca pml ucx --mca btl self --host f8n07:2,f8n08:2 ./a.out 4000
iter count is 4000
Size of MPI_COMM_WORLD = 4
0/4) Create 0:   0.03 sec (  0.00 sec per iteration)
0/4) Create 1:   0.04 sec (  0.00 sec per iteration)
0/4) Create 2:   0.23 sec (  0.00 sec per iteration)
0/4) Create 3:   1.22 sec (  0.00 sec per iteration)
0/4) Total   :   1.52 sec (  0.00 sec per iteration)
3/4) Finalize:   0.22 sec
2/4) Finalize:   0.22 sec
1/4) Finalize:   0.75 sec
0/4) Finalize:   0.75 sec
```